RefractiveIndex

RefractiveIndex.RefractiveMaterial

RefractiveIndex.RefractiveMaterial

RefractiveIndex.RefractiveMaterial — Method

RefractiveMaterial(shelf, book, page)

Load the refractive index data for the material corresponding to the specified shelf, book, and page within the refractiveindex.info database. The data can be queried by calling the returned RefractiveMaterial object at a given wavelength.

Examples

julia> MgLiTaO3 = RefractiveMaterial("other", "Mg-LiTaO3", "Moutzouris-o")
"Mg-LiTaO3 (Moutzouris et al. 2011: n(o) 0.450-1.551 µm; 8 mol.% Mg)"

julia> MgLiTaO3(0.45) # default unit is microns
2.2373000025056826

julia> using Unitful

julia> MgLiTaO3(450u"nm") # auto-conversion from generic Unitful.jl length units
2.2373000025056826

julia> MgLiTaO3(450e-9, "m") # strings can be used to specify units (parsing is cached)
2.2373000025056826

RefractiveIndex.RefractiveMaterial — Method

RefractiveMaterial(url::String)

Extracts the shelf, book, and page from a refractiveindex.info URL and loads the corresponding data from the local database (does not require an active internet connection).

Warning

The refractiveindex.info website is regularly updated and may contain materials not yet available in the local copy of the database, which is updated on a roughly annual basis. Future versions of this package may allow these new entries to be automatically downloaded on demand.

Examples

julia> Ar = RefractiveMaterial("https://refractiveindex.info/?shelf=main&book=Ar&page=Peck-15C")
"Ar (Peck and Fisher 1964: n 0.47-2.06 µm; 15 °C)"

julia> Ar(532, "nm")
1.0002679711455778